+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Changed navigation hotkeys in torrent details to 'ofpt1234', allowed navigation in tables
+
 ### Removed
 
 ## [1.1.0] - 2025-11-20 - Moon Rabbit